+++ src/configs/vitest.ts
@@ -1,0 +1,6 @@
+import type { Linter } from 'eslint';
+import { configs } from '../index.js';
+
+const vitest: Linter.Config = configs.vitest;
+
+export default vitest;